requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
opencv-contrib-python-headless==4.10.0.84
numpy==2.1.1
pytesseract==0.3.13
passporteye==2.2.1
requests==2.32.3
python-multipart==0.0.9
aiofiles==24.1.0